crypto_ai_trading/requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Database
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0

# Technical Analysis
ta>=0.10.2
pandas-ta>=0.3.14b

# Backtesting
vectorbt>=0.25.0
ccxt>=4.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging & Monitoring
tensorboard>=2.13.0
wandb>=0.15.0
tqdm>=4.65.0
colorlog>=6.7.0
python-json-logger>=2.0.0

# Model specific
einops>=0.6.1
transformers>=4.30.0

# Data handling
h5py>=3.8.0
pyarrow>=11.0.0

# Development
jupyter>=1.0.0
black>=23.0.0
pytest>=7.3.0

# System monitoring
psutil>=5.9.0

# CLI Enhancement (optional)
rich>=13.0.0